requirements.txt
fastapi
uvicorn
sqlalchemy
pydantic
<<<<<<< HEAD
python-jose
passlib
=======
passlib
textblob
nltk
streamlit
streamlit-folium
folium
requests
pandas
python-multipart
fuzzywuzzy
python-Levenshtein
>>>>>>> 8a99b5efafade8f81e0e2a501dd79446e59770e3
